+++ com.unity.netcode.gameobjects/Runtime/Components/Helpers/UnifiedBootstrap.cs
@@ -1,7 +1,9 @@
+#if UNIFIED_NETCODE && !NETCODE_GAMEOBJECT_BRIDGE_EXPERIMENTAL
+#error UNIFIED_NETCODE also requires NETCODE_GAMEOBJECT_BRIDGE_EXPERIMENTAL: Netcode for Entities keeps its GameObject bridge (GhostObject, GhostBehaviour, GhostField) internal until that define is set.
+#endif
 #if UNIFIED_NETCODE
 using System;
 using Unity.Entities;
-using Unity.NetCode;
 using UnityEngine;
 
 namespace Unity.Netcode